+ remove negtransvert flag + fix tab not being selected after adding transverter + fix update current radio config appearing multiple times in rigcontrol + test for zero removedradio tabs and transvertabs

+++ mqt/MqtRigControl/setupdialog.cpp
@@ -471,33 +471,42 @@ void SetupDialog::saveSettings()
     fileNameRadio = RADIO_PATH_LOGGER + FILENAME_AVAIL_RADIOS;
     QSettings configRadio(fileNameRadio, QSettings::IniFormat);
 
-    for (int i = 0; i < removeRadioTabs.count(); i++)
+    if (removeRadioTabs.count() > 0)
     {
-        configRadio.beginGroup(removeRadioTabs[i]);
-        configRadio.remove("");   // remove all keys for this group
-        configRadio.endGroup();
-        fileNameTransVert = TRANSVERT_PATH_LOGGER + removeRadioTabs[i] + FILENAME_TRANSVERT_RADIOS;
-        if (QFile::exists(fileNameTransVert))
+        for (int i = 0; i < removeRadioTabs.count(); i++)
         {
-            QFile::remove(fileNameTransVert);
+            configRadio.beginGroup(removeRadioTabs[i]);
+            configRadio.remove("");   // remove all keys for this group
+            configRadio.endGroup();
+            fileNameTransVert = TRANSVERT_PATH_LOGGER + removeRadioTabs[i] + FILENAME_TRANSVERT_RADIOS;
+            if (QFile::exists(fileNameTransVert))
+            {
+                QFile::remove(fileNameTransVert);
+            }
+
         }
+        removeRadioTabs.clear();
 
     }
-    removeRadioTabs.clear();
 
-    for (int i = 0; i < renameRadioTabs.count(); i++)
+    if (renameRadioTabs.count() > 0)
     {
-        configRadio.beginGroup(renameRadioTabs[i]);
-        configRadio.remove("");   // remove all keys for this group
-        configRadio.endGroup();
-        fileNameTransVert = TRANSVERT_PATH_LOGGER + renameRadioTabs[i] + FILENAME_TRANSVERT_RADIOS;
-        if (QFile::exists(fileNameTransVert))
+        for (int i = 0; i < renameRadioTabs.count(); i++)
         {
-            QFile::remove(fileNameTransVert);
+            configRadio.beginGroup(renameRadioTabs[i]);
+            configRadio.remove("");   // remove all keys for this group
+            configRadio.endGroup();
+            fileNameTransVert = TRANSVERT_PATH_LOGGER + renameRadioTabs[i] + FILENAME_TRANSVERT_RADIOS;
+            if (QFile::exists(fileNameTransVert))
+            {
+                QFile::remove(fileNameTransVert);
+            }
+
         }
+        renameRadioTabs.clear();
 
     }
-    renameRadioTabs.clear();
+
 
     bool currRadioChanged = false;
 
@@ -532,23 +541,31 @@ void SetupDialog::saveSettings()
             QSettings  configTransVert(fileNameTransVert, QSettings::IniFormat);
 
             radioTab[i]->addedTransVertTabs.clear();
-            for (int t = 0; t < radioTab[i]->removedTransVertTabs.count(); t++)
-            {
-                QSettings config(fileNameTransVert, QSettings::IniFormat);
-                config.beginGroup(radioTab[i]->removedTransVertTabs[t]);
-                config.remove("");      // remove all keys for this group
-                config.endGroup();
-            }
-            radioTab[i]->removedTransVertTabs.clear();
 
-            for (int t = 0; t < radioTab[i]->renamedTransVertTabs.count(); t++)
+            if (radioTab[i]->removedTransVertTabs.count() > 0)
             {
-                QSettings config(fileNameTransVert, QSettings::IniFormat);
-                config.beginGroup(radioTab[i]->renamedTransVertTabs[t]);
-                config.remove("");      // remove all keys for this group
-                config.endGroup();
+                for (int t = 0; t < radioTab[i]->removedTransVertTabs.count(); t++)
+                {
+                    QSettings config(fileNameTransVert, QSettings::IniFormat);
+                    config.beginGroup(radioTab[i]->removedTransVertTabs[t]);
+                    config.remove("");      // remove all keys for this group
+                    config.endGroup();
+                }
+                radioTab[i]->removedTransVertTabs.clear();
             }
-            radioTab[i]->renamedTransVertTabs.clear();
+
+            if (radioTab[i]->renamedTransVertTabs.count() > 0)
+            {
+                for (int t = 0; t < radioTab[i]->renamedTransVertTabs.count(); t++)
+                {
+                    QSettings config(fileNameTransVert, QSettings::IniFormat);
+                    config.beginGroup(radioTab[i]->renamedTransVertTabs[t]);
+                    config.remove("");      // remove all keys for this group
+                    config.endGroup();
+                }
+                radioTab[i]->renamedTransVertTabs.clear();
+            }
+
 
             if (radioTab[i]->getRadioData()->numTransverters > 0)
             {
@@ -573,26 +590,14 @@ void SetupDialog::saveSettings()
 
                         radioTab[i]->transVertTab[t]->transVertValueChanged = false;
                     }
-
-
                 }
-
-
             }
-
-
-
-
-
-        }
-
-        if (currRadioChanged)
-        {
-            emit currentRadioSettingChanged(radioTab[i]->getRadioData()->radioName);
         }
         radioTab[i]->radioValueChanged = false;
-
-
+    }
+    if (currRadioChanged)
+    {
+        emit currentRadioSettingChanged(currentRadioName);
     }
 }
 
@@ -715,7 +720,6 @@ void SetupDialog::saveTranVerterSetting(int radioNum, int transVertNum, QSetting
     config.setValue("targetFreq", radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->targetFreq);
     config.setValue("offsetString", radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->transVertOffsetStr);
     config.setValue("offsetDouble", radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->transVertOffset);
-    config.setValue("negOffset", radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->transVertNegative);
     config.setValue("antSwNumber", radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->antSwitchNum);
     config.setValue("enableTransVertSw", radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->enableTransSwitch);
     config.setValue("transVertSw", radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->transSwitchNum);
@@ -737,7 +741,6 @@ void SetupDialog::readTranVerterSetting(int radioNum, int transVertNum, QSetting
     radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->targetFreq = config.value("targetFreq", 0.0).toDouble();
     radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->transVertOffsetStr = config.value("offsetString", "00.000.000.000").toString();
     radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->transVertOffset = config.value("offsetDouble", 0.0).toDouble();
-    radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->transVertNegative = config.value("negOffset", false).toBool();
     radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->antSwitchNum = config.value("antSwNumber", "0").toString();
     radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->enableTransSwitch = config.value("enableTransVertSw", false).toBool();
     radioTab[radioNum]->getRadioData()->transVertSettings[transVertNum]->transSwitchNum = config.value("transVertSw", "0").toString();
